only permission admin

+++ controllers/categoryCollectionControllers.js
@@ -3,6 +3,9 @@ const { upload } = require("../utils/file");
 
 const categoryCreate = async (req, res, next) => {
     try {
+        if(req.user?.isAdmin !== true){
+            return res.status(400).json({error:{permission:"You can perform only Admin"}})
+        }
         const category = req.body?.category;
         let img = req.body?.img;
         const age = req.body?.age;
@@ -23,6 +26,9 @@ const categoryCreate = async (req, res, next) => {
 }
 const subCategoryCreate = async (req, res, next) => {
     try {
+        if(req.user?.isAdmin !== true){
+            return res.status(400).json({error:{permission:"You can perform only Admin"}})
+        }
         const category = req.body?.category;
         const subCategory = req.body?.subCategory;
         let img = req.body?.img;
@@ -44,6 +50,9 @@ const subCategoryCreate = async (req, res, next) => {
 }
 const insideSubCategoryCreate = async (req, res, next) => {
     try {
+        if(req.user?.isAdmin !== true){
+            return res.status(400).json({error:{permission:"You can perform only Admin"}})
+        }
         const category = req.body?.category;
         const subCategory = req.body?.subCategory;
         const insideSubCategory = req.body?.insideSubCategory;
@@ -60,6 +69,9 @@ const insideSubCategoryCreate = async (req, res, next) => {
 }
 const insidePackTypeCreate = async (req, res, next) => {
     try {
+        if(req.user?.isAdmin !== true){
+            return res.status(400).json({error:{permission:"You can perform only Admin"}})
+        }
         const packType = req.body?.packType;
         const category = req.body?.category;
         const subCategory = req.body?.subCategory;
@@ -77,6 +89,9 @@ const insidePackTypeCreate = async (req, res, next) => {
 }
 const insideServingSizeCreate = async (req, res, next) => {
     try {
+        if(req.user?.isAdmin !== true){
+            return res.status(400).json({error:{permission:"You can perform only Admin"}})
+        }
         const servingSize = req.body?.servingSize;
         const category = req.body?.category;
         const subCategory = req.body?.subCategory;
@@ -93,6 +108,9 @@ const insideServingSizeCreate = async (req, res, next) => {
 }
 const categoryGet = async (req, res, next) => {
     try {
+        if(req.user?.isAdmin !== true){
+            return res.status(400).json({error:{permission:"You can perform only Admin"}})
+        }
         let { page = 1, limit = 30 } = req.query;
         limit = parseInt(limit);
         const keyword = req.query.search ? {
@@ -110,6 +128,9 @@ const categoryGet = async (req, res, next) => {
 }
 const subCategoryGet = async (req, res, next) => {
     try {
+        if(req.user?.isAdmin !== true){
+            return res.status(400).json({error:{permission:"You can perform only Admin"}})
+        }
         let { page = 1, limit = 30 } = req.query;
         limit = parseInt(limit);
         let keyword;
@@ -137,6 +158,9 @@ const subCategoryGet = async (req, res, next) => {
 }
 const insideSubCategoryGet = async (req, res, next) => {
     try {
+        if(req.user?.isAdmin !== true){
+            return res.status(400).json({error:{permission:"You can perform only Admin"}})
+        }
         let { page = 1, limit = 30 } = req.query;
         limit = parseInt(limit);
         let keyword;
@@ -163,6 +187,9 @@ const insideSubCategoryGet = async (req, res, next) => {
 }
 const insidePackTypeGet = async (req, res, next) => {
     try {
+        if(req.user?.isAdmin !== true){
+            return res.status(400).json({error:{permission:"You can perform only Admin"}})
+        }
         let { page = 1, limit = 30 } = req.query;
         limit = parseInt(limit);
         let keyword;
@@ -190,6 +217,9 @@ const insidePackTypeGet = async (req, res, next) => {
 }
 const insideServingSizeGet = async (req, res, next) => {
     try {
+        if(req.user?.isAdmin !== true){
+            return res.status(400).json({error:{permission:"You can perform only Admin"}})
+        }
         let { page = 1, limit = 30 } = req.query;
         limit = parseInt(limit);
         let keyword;
@@ -218,6 +248,9 @@ const insideServingSizeGet = async (req, res, next) => {
 
 const categoryUpdate = async (req, res, next) => {
     try {
+        if(req.user?.isAdmin !== true){
+            return res.status(400).json({error:{permission:"You can perform only Admin"}})
+        }
         const category = req.body?.category;
         const img = req.body?.img;
         const age = req.body?.age;
@@ -238,6 +271,9 @@ const categoryUpdate = async (req, res, next) => {
 }
 const subCategoryUpdate = async (req, res, next) => {
     try {
+        if(req.user?.isAdmin !== true){
+            return res.status(400).json({error:{permission:"You can perform only Admin"}})
+        }
         const subCategory = req.body?.subCategory;
         const img = req.body?.img;
         const age = req.body?.age;
@@ -258,6 +294,9 @@ const subCategoryUpdate = async (req, res, next) => {
 }
 const insideSubCategoryUpdate = async (req, res, next) => {
     try {
+        if(req.user?.isAdmin !== true){
+            return res.status(400).json({error:{permission:"You can perform only Admin"}})
+        }
         const insideSubCategory = req.body?.insideSubCategory;
         const update = await InsideSubCategory.findOneAndUpdate({ _id: req.params.id }, {
             insideSubCategory
@@ -274,6 +313,9 @@ const insideSubCategoryUpdate = async (req, res, next) => {
 }
 const insidePackTypeUpdate = async (req, res, next) => {
     try {
+        if(req.user?.isAdmin !== true){
+            return res.status(400).json({error:{permission:"You can perform only Admin"}})
+        }
         const packType = req.body?.packType;
         const update = await InsidePackType.findOneAndUpdate({ _id: req.params.id }, {
             packType
@@ -291,6 +333,9 @@ const insidePackTypeUpdate = async (req, res, next) => {
 }
 const insideServingSizeUpdate = async (req, res, next) => {
     try {
+        if(req.user?.isAdmin !== true){
+            return res.status(400).json({error:{permission:"You can perform only Admin"}})
+        }
         const servingSize = req.body?.servingSize;
         const update = await InsideServingSize.findOneAndUpdate({ _id: req.params.id }, {
             servingSize
@@ -309,6 +354,9 @@ const insideServingSizeUpdate = async (req, res, next) => {
 //delete Category 
 const deleteCategory = async (req, res, next) => {
     try {
+        if(req.user?.isAdmin !== true){
+            return res.status(400).json({error:{permission:"You can perform only Admin"}})
+        }
         const deleted = Category.findOneAndRemove({ _id: req.params.id });
         if (!deleted) {
             res.status.json({ error: { "exist": "Category Delete Failed! Not Exists!" } })
@@ -323,6 +371,9 @@ const deleteCategory = async (req, res, next) => {
 }
 const deleteSubCategory = async (req, res, next) => {
     try {
+        if(req.user?.isAdmin !== true){
+            return res.status(400).json({error:{permission:"You can perform only Admin"}})
+        }
         const deleted = SubCategory.findOneAndRemove({ _id: req.params.id });
         if (!deleted) {
             res.status.json({ error: { "exist": "Sub Category Delete Failed! Not Exist!" } })
@@ -337,6 +388,9 @@ const deleteSubCategory = async (req, res, next) => {
 }
 const deleteInsideSubCategory = async (req, res, next) => {
     try {
+        if(req.user?.isAdmin !== true){
+            return res.status(400).json({error:{permission:"You can perform only Admin"}})
+        }
         const deleted = InsideSubCategory.findOneAndRemove({ _id: req.params.id });
         if (!deleted) {
             res.status.json({ error: { "exist": "Inside Sub Category Delete Failed! Not Exists!" } })
@@ -351,6 +405,9 @@ const deleteInsideSubCategory = async (req, res, next) => {
 }
 const deletePackType = async (req, res, next) => {
     try {
+        if(req.user?.isAdmin !== true){
+            return res.status(400).json({error:{permission:"You can perform only Admin"}})
+        }
         const deleted = InsidePackType.findOneAndRemove({ _id: req.params.id });
         if (!deleted) {
             res.status.json({ error: { "exist": "Pack Type Delete Failed! Not Exist!" } })
@@ -365,6 +422,9 @@ const deletePackType = async (req, res, next) => {
 }
 const deleteServingSize = async (req, res, next) => {
     try {
+        if(req.user?.isAdmin !== true){
+            return res.status(400).json({error:{permission:"You can perform only Admin"}})
+        }
         const deleted = InsideServingSize.findOneAndRemove({ _id: req.params.id });
         if (!deleted) {
             res.status.json({ error: { "exist": "Surving Size Delete Failed! Not Exist" } })
